rats/rvp: Keep reference count of RVP entries
To avoid duplication of registered reference values, keep a reference
count of all used entries. Calculate a SHA256 hash of reference values
when they are submitted, and update the entry's reference count if an
existing one is found. If an existing entry is not found, add a new
entry into the RVP map.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>